% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_gibbs.R
\name{threshold_gibbs}
\alias{threshold_gibbs}
\title{threshold_gibbs}
\usage{
threshold_gibbs(
  data1,
  Y,
  motif_len,
  dict,
  burn_in = 2000,
  end_times = 3000,
  result_path,
  only_get_func = F,
  input_prior = F,
  prior = 0
)
}
\arguments{
\item{data1}{an array whose element is a string, e.g, ('ATGC','TTGG')}

\item{Y}{the binding scores}

\item{motif_len}{the length of the motif}

\item{dict}{named vector for letters}

\item{burn_in}{the burn_in setting for Gibbs}

\item{end_times}{the total iterations}

\item{result_path}{where to store the results}

\item{only_get_func}{whether only loading the functions}

\item{input_prior}{whether inputting priors by users}

\item{prior}{different prior settings}
}
\value{
a dataframe with three columns: pep, real_score, est_score
}
\description{
this is the main function of MHCpBT, input sequences and corresponding binding scores
}
